first try with gltf animation (rotations)
rotations are calculated as the inverse global pose matrix of the parent
bone multiplied by the pose global matrix of the bone itself.

When size is not decimeter animation is still distorted, there is no
real animation, at least for one frame I got a distorted character (obviously the
weights are miscalculated or direction y-up is still terribly wrong.