file-system help added and fix of #6
The asset-download does not supply litspheres afaik (not needed in
Blender), so I ignored the errors not to discourage people atm. I simply
switch to phong-shading (and changed the error message a bit).

In menu help + file-system shows the place, where to put them. Mine are
all in user-space. I should link that png into readme as well.