accept reduced skeletons
without references like in default-no-toes skeleton, the original name of
the bone is searched. This skeleton always created an animation without
any movement at all :(

the khronos gltf-validator complained about small values potentially
created by calculation with numpy linalg(?). So the last column of that
matrix is now set to 0, 0, 0, 1