offscreen render with transparent background changed
At least for Linux now the offscreen render does no longer create a
transparent character with transparent clothes on a transparent
background. I kept the alpha values after the first render and if they
are greater than the final image, they are replaced by the values of the
body. Do not ask me, if there is a better solution :(