OSX support started
Although OSX cannot deal with openGL 3.3, we managed to add different
shaders for this OS which start with v12. If internally the environment
(osindex) is set to 2 these shaders are used. As long as we are not able
to use e.g. Vulkan, this is the best way to do it. Apple seems to
initialize the GUI after the rest is drawn. Therefore I preset blocking
of openGL to true. These are the first steps, it could be, that is does
not yet work.